Move stream fixtures up a level
The stream bits are useful for more than just ATCP.
